% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconmiR.R
\name{DeconmiR}
\alias{DeconmiR}
\title{Cell type deconvolution from miRNA expression profiling}
\usage{
DeconmiR(
  data.m,
  ref.m,
  method = c("RPC", "SVR", "CP", "NNLS", "SLE"),
  maxit = 50,
  nu.v = c(0.25, 0.5, 0.75),
  constraint = c("inequality", "equality")
)
}
\arguments{
\item{data.m}{miRNA Expression profile: A data matrix with rows labeling the
miRNAs (should use miRBase ID as in ref.m) and columns labeling samples
(e.g. primary tumor specimens). Missing value is not allowed and all values
should be log-transformed.}

\item{ref.m}{Reference profile: The signature matrix should be a miRNAs (rows)
by cell types (columns) matrix containing log-normalized expression
values of signature miRNAs.}

\item{method}{Deconvolution method: Choice of the deconvolution methods to
be used: ("RPC", "SVR", "CP", "NNLS", "SLE"). By default is the "RPC".}

\item{maxit}{The limit of the number of IWLS iterations, only used for "RPC" mode.}

\item{nu.v}{A vector of several candidate nu values, which is needed for
nu-classification, nu-regression, and one-classification in svm, only used
for "SVR" mode. The best estimation results among all candidate nu will be
automatically returned.}

\item{constraint}{Choice of either 'inequality' or 'equality' normalization
constraint, only used in "CP" mode, By default is 'inequality' (i.e sum of
weights adds to a number less or equal than 1), which was implemented in
Houseman et al (2012).}
}
\value{
A list with the following entries: estF: a matrix of the estimated
fractions, rows are samples and columns are cell types; ref: the reference
matrix used; dataREF: the subset of the input data matrix with only the
miRNAs defined in the reference matrix.
}
\description{
Cell type proportions prediction from miRNA expression profiling
data by applying multiple deconvolution methods. Inference proceeds via one
of 5 methods (Robust Partial Correlations-RPC, Support Vector Regression-SVR,
Constrained Projection-CP, Non-Negative Least Squares-NNLS, Solution of Linear
Equations-SLE), as determined by the user.
}
\examples{
data(miRExpCount_LAML.m)
data(refMatrix_Blood1.m)
miRexpRPM_LAML.m <- getRPM(miRExpCount_LAML.m)
miRexpNorm_LAML.m <- logNormalize(miRexpRPM_LAML.m)
est.o <- DeconmiR(miRexpNorm_LAML.m, refMatrix_Blood1.m, method = 'RPC')
cellfrac.m <- est.o$estF

}
