% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenericMatrix}
\alias{GenericMatrix}
\title{Generic reference matrices for solid tissues}
\format{
A list with 2 different reference matrices
}
\usage{
data(GenericMatrix)
}
\description{
Generic reference matrices consists of two references used for the
Hierarchical mode of DeconmiR (hDeconmiR), which is used for cell type deconvolution
of any solid tissue. The first is the reference centroids for epithelial cells,
fibroblasts and total immune cells, and the second is the reference for 7 subtypes of
immune cells, such as B-cells, NK cells, monocytes and etc.
\itemize{
\item ref1
\item ref2
}
}
\keyword{datasets}
