% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNormalize.R
\name{logNormalize}
\alias{logNormalize}
\title{Title Log-transformation and normalization of miRNA expression}
\usage{
logNormalize(miRExp.m, quantile = FALSE)
}
\arguments{
\item{miRExp.m}{miRNA expression table: a miRNAs (rows) by samples (columns) matrix.
Either RPM value or microarray based expression value.}

\item{quantile}{Choice of either 'TRUE' or 'FALSE' for quantile normalization.
We suggest use normalization for microarray based expression.}
}
\value{
miRNA expression table with log2 transformed and normalized: a miRNAs
(rows) by samples (columns) matrix.
}
\description{
Log-transform and further normalize the miRNA expression, which is
suitable normalization for deconvolution.
}
\examples{
data(miRExpCount_LAML.m)
miRexpRPM_LAML.m <- getRPM(miRExpCount_LAML.m)
miRexpNorm_LAML.m <- logNormalize(miRexpRPM_LAML.m)

}
