% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRExpCount_BRCA.m}
\alias{miRExpCount_BRCA.m}
\title{miRNA expression profiles (TCGA BRCA)}
\format{
A matrix with 2235 rows (miRNAs) and 12 variables (samples)
}
\usage{
data(miRExpCount_BRCA.m)
}
\description{
An example data table that containing the miRNA read count from
bulk miRNA-seq of 12 Breast tissue samples form TCGA BRCA (Breast Invasive Carcinoma)
cohort. Only used for demonstration purpose.
}
\keyword{datasets}
