% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRExpCount_LAML.m}
\alias{miRExpCount_LAML.m}
\title{miRNA expression profiles (TCGA LAML)}
\format{
A matrix with 1807 rows (miRNAs) and 12 variables (samples)
}
\usage{
data(miRExpCount_LAML.m)
}
\description{
An example data table that containing the miRNA read count from
bulk miRNA-seq of 12 Whole Blood samples form TCGA LAML (Acute Myeloid Leukemia)
cohort. Only used for demonstration purpose.
}
\keyword{datasets}
