% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refMatrix_Blood1.m}
\alias{refMatrix_Blood1.m}
\title{Whole blood reference matrix of 145 miRNAs and 6 blood cell subtypes}
\format{
A matrix with 145 rows (miRNAs) and 6 variables (Cell Types)
}
\usage{
data(refMatrix_Blood1.m)
}
\description{
Reference-based cell-type fraction estimation methods rely on a
prior defined reference matrix. We used miRNA expression profiling data from
purified blood cell types from Simonas et al (2012) to construct this whole
blood reference matrix by using DEM methods. It contains 145 miRNAs of 6
blood cell subtypes:
\itemize{
\item CD4+ T-cells
\item CD8+ T-cells
\item B-cells
\item NK-cells
\item Monocytes
\item Neutrophils
}
}
\keyword{datasets}
