% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refMatrix_Blood2.m}
\alias{refMatrix_Blood2.m}
\title{Whole blood reference matrix of 140 miRNAs and 6 blood cell subtypes}
\format{
A matrix with 140 rows (miRNAs) and 6 variables (Cell Types)
}
\usage{
data(refMatrix_Blood2.m)
}
\description{
Reference-based cell-type fraction estimation methods rely on a
prior defined reference matrix. We used miRNA expression profiling data from
purified blood cell types from FANTOM5 project to construct this whole
blood reference matrix by using DEM method. It contains 140 miRNAs of 6
blood cell subtypes:
\itemize{
\item CD4+ T-cells
\item CD8+ T-cells
\item B-cells
\item NK-cells
\item Monocytes
\item Neutrophils
}
}
\keyword{datasets}
