% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refMatrix_Breast.m}
\alias{refMatrix_Breast.m}
\title{Breast tissue reference matrix of 271 miRNAs and 10 cell subtypes}
\format{
A matrix with 271 rows (miRNAs) and 10 variables (Cell Types)
}
\usage{
data(refMatrix_Breast.m)
}
\description{
Reference-based cell-type fraction estimation methods rely on a
prior defined reference matrix. We used miRNA expression profiling data from
purified blood cell types from FANTOM5 to construct this whole blood reference
matrix. It contains 271 miRNAs of 6 blood cell subtypes:
\itemize{
\item Epithelial
\item Fibroblast
\item Adipocyte
\item CD4+ T-cells
\item CD8+ T-cells
\item B-cells
\item NK-cells
\item Monocytes
\item Macrophage
\item Neutrophils
}
}
\keyword{datasets}
