% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SolidTissueMatrix}
\alias{SolidTissueMatrix}
\title{Reference matrices for 9 different tissues}
\format{
A list with 9 different reference matrices
}
\usage{
data(SolidTissueMatrix)
}
\description{
We used miRNA expression profiling data from purified cells
obtained from FANTOM5 to construct reference for solid tissues including the
Breast, Lung, Liver, Kidney, Prostate, Esophagus, Respiratory_tract, Vessel
and Skin. For breast tissue, it contains 10 cell subtypes dominated by the
epithelial, fibroblast, adipocyte and immune-cell component of seven cell types.
whereas for other tissues, the reference mainly contains 9 cell types that
dominated by the epithelial, fibroblast and immune-cell components.
\itemize{
\item Breast
\item Lung
\item Liver
\item Kidney
\item Prostate
\item Esophagus
\item Respiratory_tract
\item Vessel
\item Skin
}
}
\keyword{datasets}
